Refactored Checkbox
